+++ src/main/java/com/guildboard/Model/repository/AssignmentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.guildboard.Model.entity.Assignment;
 
-public interface AssignmentRepository extends JpaRepository<Assignment, Integer> {
+public interface AssignmentRepository extends JpaRepository<Assignment, Long> {
     
 }
