+++ src/main/java/com/guildboard/Model/repository/QuestRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.guildboard.Model.entity.Quest;
 
-public interface QuestRepository extends JpaRepository<Quest, Integer> {
+public interface QuestRepository extends JpaRepository<Quest, Long> {
 
     Quest findById(int id);
     Quest findBytitle(String title);
